% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coef_traceplot}
\alias{coef_traceplot}
\title{Draw a coefficient traceplot}
\usage{
coef_traceplot(coef.mat, varnames = NULL)
}
\arguments{
\item{coef.mat}{The matrix of coefficients (one coefficient vector per row).}

\item{varnames}{(Optional) list of variable name labels to make the traceplot more readable.}
}
\description{
This function draws a 'traceplot' of coefficient values vs. number of iterations.
}

