% Generated by roxygen2 (4.0.0): do not edit by hand
\name{eeboost}
\alias{eeboost}
\title{EEBoost}
\usage{
eeboost(...)
}
\arguments{
\item{...}{Arguments to \code{\link{threeboost}}.}
}
\description{
Alias for ThrEEBoost (which defaults to a threshold value of 1).
}
\seealso{
\code{\link{threeboost}}
}

