% Generated by roxygen2 (4.0.0): do not edit by hand
\name{QIC}
\alias{QIC}
\title{Pan's QIC}
\usage{
QIC(Y, X, b, family = "gaussian")
}
\arguments{
\item{Y}{A vector of outcomes.}

\item{X}{A matrix of predictors.}

\item{b}{A vector of regression coefficients (e.g., a row from the coefficient matrix produced by \code{\link{geeboost}})}

\item{family}{Version of QIC to implement, for either \code{"gaussian"}, \code{"binomial"} or \code{"poisson"} outcomes.
Should match the \code{family} argument used in the original boosting algorithm.}
}
\description{
Calculates a simple version of Pan's QIC for a GEE model defined by
 a vector of regression coefficients.
}

