% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ee.GEE}
\alias{ee.GEE}
\alias{ee.GEE.aux}
\title{GEE estimating functions}
\usage{
ee.GEE(Y, X, b, mu.Y = function(eta) {     eta }, v.Y = function(eta) {    
  rep(1, length(eta)) }, aux = NULL, id = 1:length(Y),
  uid = sort(unique(id)), rows.indivs = lapply(uid, function(j) {    
  which(id == j) }), corstr = "ind")

ee.GEE.aux(Y, X, b, mu.Y = function(eta) {     eta }, v.Y = function(eta) { 
     rep(1, length(eta)) }, id = 1:length(Y), uid = sort(unique(id)),
  rows.indivs = lapply(uid, function(j) {     which(id == j) }))
}
\arguments{
\item{Y}{Vector of (correlated) outcomes}

\item{X}{Matrix of predictors}

\item{b}{Vector of coefficients}

\item{mu.Y}{Mean function}

\item{v.Y}{Variance function}

\item{aux}{Auxiliary function for coputing (co)variance parameters}

\item{id}{Vector of cluster IDs}

\item{uid}{Vector of unique subject IDs}

\item{rows.indivs}{List of rows of \code{X} corresponding to each subject ID}

\item{corstr}{Working correlation structure}
}
\description{
Internal functions for computing the GEE. Should generally not be called by user.
}
\section{Functions}{
\itemize{
\item \code{}: 
}}

